LayoutTests/media/video-error-does-not-exist.html

<video src=content/does-not-exist.mpeg></video>
<script src=video-test.js></script>
<script>
test("!video.error");
waitForEventAndTest("load", "false");
waitForEventTestAndEnd("emptied", "video.error.code == MediaError.MEDIA_ERR_NETWORK");
</script>
