LayoutTests/media/video-error-does-not-exist-expected.txt

TEST(!video.error) OK
EVENT(emptied) TEST(video.error.code == MediaError.MEDIA_ERR_NETWORK) OK
END OF TEST
